Finished insert method

+++ app/src/main/java/com/csabafarkas/popularmovies/models/Movie.java
@@ -31,7 +31,7 @@ public class Movie implements Parcelable, PopularMoviesModel
     private String homepage;
     @SerializedName("id")
     @Expose
-    private Integer id;
+    private Long id;
     @SerializedName("imdb_id")
     @Expose
     private String imdbId;
@@ -112,7 +112,7 @@ public class Movie implements Parcelable, PopularMoviesModel
         this.budget = ((Integer) in.readValue((Integer.class.getClassLoader())));
         in.readList(this.genres, (com.csabafarkas.popularmovies.models.Genre.class.getClassLoader()));
         this.homepage = ((String) in.readValue((String.class.getClassLoader())));
-        this.id = ((Integer) in.readValue((Integer.class.getClassLoader())));
+        this.id = ((Long) in.readValue((Long.class.getClassLoader())));
         this.imdbId = ((String) in.readValue((String.class.getClassLoader())));
         this.originalLanguage = ((String) in.readValue((String.class.getClassLoader())));
         this.originalTitle = ((String) in.readValue((String.class.getClassLoader())));
@@ -184,11 +184,11 @@ public class Movie implements Parcelable, PopularMoviesModel
         this.homepage = homepage;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
